requirements.txt
streamlit

pymongo

python-dotenv
flask-cors
bcrypt
flask-jwt-extended
flask-marshmallow
flask-limiter
pymongo[srv]
python-dotenv